backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Task Queue
celery==5.3.6
redis==5.0.1

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Web Scraping (from existing core modules)
beautifulsoup4==4.12.3
lxml==5.1.0

# OpenAI
openai==1.10.0

# WebSocket
python-socketio==5.11.0

# Utilities
python-slugify==8.0.1
pytz==2024.1
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
